Split input and grid validation domains
